requirements.txt
streamlit>=1.31.0
fastapi>=0.109.0
uvicorn>=0.27.0
pillow>=10.2.0
torch>=2.2.0
transformers>=4.37.0
pandas>=2.2.0
scikit-learn>=1.4.0
python-multipart>=0.0.9
requests>=2.31.0
numpy>=1.26.0
nltk>=3.8.1
rank-bm25>=0.2.2
bottleneck>=1.3.6
datasets==2.17.1
python-dotenv==1.0.0
opencv-python-headless==4.9.0.80
google-cloud-storage==2.14.0
firebase-admin==6.4.0
diffusers==0.25.0
accelerate==0.27.0
huggingface-hub==0.20.3
safetensors==0.4.2
pydantic==2.6.1
starlette==0.36.3
openai==1.12.0
plotly==5.19.0 